requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-core==1.29.5
azure-storage-blob==12.18.3
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.1
fuzzywuzzy==0.18.0
h11==0.14.0
idna==3.4
isodate==0.6.1
joblib==1.3.2
Levenshtein==0.23.0
numpy==1.26.1
pandas==2.1.1
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==3.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-Levenshtein==0.23.0
python-multipart==0.0.6
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.23.2
